pagination done

+++ Services/ServiceClasses/StudentService.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.EntityFrameworkCore.Metadata.Internal;
+﻿using Microsoft.EntityFrameworkCore;
 using University_Information_System.Data;
 using University_Information_System.Models;
 using University_Information_System.Services.ServiceInterfaces;
@@ -15,46 +15,51 @@ namespace University_Information_System.Services.ServiceClasses
             this.db = db;
         }
 
-        public void AddStudent(Student student)
+        public async Task AddStudent(Student student)
         {
-            db.Student.Add(student);
-            db.SaveChanges();
+           await db.Student.AddAsync(student);
+           await db.SaveChangesAsync();
         }
 
-        public IQueryable<Student> getAllStudent()
+        public async Task<List<Student>> getAllStudent()
         {
-            return db.Student; 
+            return await db.Student.ToListAsync(); 
         }
 
-        public void UpdateStudent(Student student)
+        public async Task UpdateStudent(Student student)
         {
             db.Student.Update(student);
-            db.SaveChanges();
+            await db.SaveChangesAsync();
         }
 
-        public Student GetStudentById(int id)
+        public async Task<Student> GetStudentById(int id)
         {
-
-            return db.Student.Find(id);
+            return await db.Student.FindAsync(id);
         }
 
-        public void DeleteStudent(Student student)
+        public async Task DeleteStudent(Student student)
         {
-            var subStudentVar = db.SubjectStudentMapped.Where(ss => ss.studentId == student.id );
+            var subStudentVar = await db.SubjectStudentMapped.Where(ss => ss.studentId == student.id).ToListAsync();
             
             db.SubjectStudentMapped.RemoveRange(subStudentVar);
-            db.SaveChanges();
+            await db.SaveChangesAsync();
             db.Student.Remove(student);
-            db.SaveChanges();
+            await db.SaveChangesAsync();
         }
 
 
 
-        public Student GetStudentDetailsById(int id)
+        public async Task<Student> GetStudentDetailsById(int id)
         {
-            var student = db.Student.Find(id);
-            student.DeptName =db.Depertment.Find(student.DepertmentId).DeptName;
-            student.Subjects = GetSubjectByStudentId(id);
+            var student = await db.Student.FindAsync(id);
+            if(student == null)return new Student();
+
+            var dept = await db.Depertment.FindAsync(student.DepertmentId);
+
+            if(dept == null)return new Student();
+
+            student.DeptName = dept.DeptName;
+            student.Subjects = await GetSubjectByStudentId(id);
 
             return student;
         }
@@ -64,29 +69,32 @@ namespace University_Information_System.Services.ServiceClasses
 
 
 
-        public void AddSubjectStudentMapped(SubjectStudentMapped subjectStudentMapped)
+        public async Task AddSubjectStudentMapped(SubjectStudentMapped subjectStudentMapped)
         {
-            db.SubjectStudentMapped.Add(subjectStudentMapped);
-            db.SaveChanges();
+            await db.SubjectStudentMapped.AddAsync(subjectStudentMapped);
+            await db.SaveChangesAsync();
         }
 
-        public void DeleteEnrolmentFromSubjectStudentMapped(int subjectId, int studentId)
+        public async Task DeleteEnrolmentFromSubjectStudentMapped(int subjectId, int studentId)
         {
-            var subStudentVar = db.SubjectStudentMapped.FirstOrDefault(ss=>ss.studentId == studentId && ss.subjectId==subjectId);
-
-            db.SubjectStudentMapped.Remove(subStudentVar);
-            db.SaveChanges();
+            var subStudentVar = await db.SubjectStudentMapped.FirstOrDefaultAsync(ss=>ss.studentId == studentId && ss.subjectId==subjectId);
+            if (subStudentVar != null)
+            {
+                db.SubjectStudentMapped.Remove(subStudentVar);
+                await db.SaveChangesAsync();
+            }
         }
 
 
 
-        public List<Subject> GetSubjectByStudentId(int id)
+        public async Task<List<Subject>> GetSubjectByStudentId(int id)
         {
-            var subjectIdOfTheStudent = db.SubjectStudentMapped.Where(ss => ss.studentId == id).Select(ss=>ss.subjectId);
-            var allSubjectOfTheStudent = db.Subject.Where(subject => subjectIdOfTheStudent.Contains(subject.id)).ToList();
+            var subjectIdOfTheStudent = await db.SubjectStudentMapped.Where(ss => ss.studentId == id).Select(ss => ss.subjectId).ToListAsync();
+            var allSubjectOfTheStudent = await db.Subject.Where(subject => subjectIdOfTheStudent.Contains(subject.id)).ToListAsync();
 
             return allSubjectOfTheStudent;
         }
+
 
     }
 }
